pyproject.toml
[project]
name = "maze"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy",
    "Pillow",
    "imageio[pyav]",
    "tqdm",
]
